refactor(agent_cmd): collapse telegram/discord into parameterized platform.py
RC: telegram.py and discord.py were 100% copy-pasted — only the platform
literal differed. make_platform_app(platform) factories the full Typer
sub-app; shims keep backward-compatible module names.

Closes #1665
